changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11814 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Pci/Mac/ORMacModel.h
@@ -63,8 +63,8 @@
 - (id) getFireWireInterface:(NSUInteger) aVenderID;
 
 #pragma mark ¥¥¥USB
-- (unsigned) usbDeviceCount;
-- (id)		 usbDeviceAtIndex:(unsigned)index;
+- (NSUInteger) usbDeviceCount;
+- (id)		 usbDeviceAtIndex:(NSUInteger)index;
 - (void)	 objectsAdded:(NSNotification*)aNote;
 - (void)	 objectsRemoved:(NSNotification*)aNote;
 - (id)		 getUSBController;
